Pedantic warnings (almost) fixed.
Common 'mistakes':
- Extra semi-colon somewhere;
- Variable Length Arrays (VLAs) converted to std::vector and
  boost:multi_array.
  If v is a std::vector<T>, &v[0] is T*, so you can pass v keep
  consistency with MPI and Epetra calls;
- Unused variables.

+++ lifev/core/array/VectorEpetraStructured.cpp
@@ -134,7 +134,7 @@ setBlockStructure( const mapVector_type& mapVector)
 		totalSize+= M_blockSize[i];
 	}
 
-    ASSERT( totalSize == this->size()," Incompatible block structure (global size does not match) ");
+    ASSERT( totalSize == static_cast<UInt>(this->size())," Incompatible block structure (global size does not match) ");
 }
 
 // ===================================================
